refactor: moved hooks to appropriate location
